Session recycling for WiFi should update the config (#1284)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1284

When a WIFI session creation requested AND the IMSI already has a session, we recycle the existing session. We should override the existing session config with the new config.
Ex. If the user switches AP, that will trigger a new CreateSession and our current implementation will recycle the old session. But since the old config still points to the previous AP, when the old AP gets around to sending a Terminate Request, the session incorrectly gets terminated.

This should fix that problem, as `terminate_subscriber` in local enforcer only terminates the subscriber that matches the given APN and IMSI.

Reviewed By: emakeev

Differential Revision: D20267498

fbshipit-source-id: c30de9aa5bf8334f0f2736398b5a58ef4cb7ef2a