Improve Grafana error messaging
Summary:
Since there are quite a few places that the Grafana sync could fail, it's important to have detailed logs of what happened and what went wrong so network admins can debug the error.

This diff adds the logging and an error display for the user when an error occurs. Additionally, it makes the sync job more robust and correct by not making any assumptions about the state of Grafana. E.g. Even if a user exists, it will still check that the user exists in the correct organization and fixes it if they don't.

I would like to make a real react component for this, but the way the App is built right now makes that pretty difficult. Rendering on the server is a good stopgap for now.

Reviewed By: rckclmbr

Differential Revision: D20144496

fbshipit-source-id: 7a63a1a14de57d99691a04c0f7f054bf58443a05