Update directoryd service iface in preparation for session ID indexer.
Summary:
## Tl;dr
We desire a non-hacky way to translate session IDs to IMSIs at the controller. This diff exposes session ID and IMSI getters/setters via the directoryd service.

## Context
The FeG sends abort-session requests to the controller, which need to get translated to an IMSI. We don't have a good way of doing this right now, so the FeG sends the session ID as a concatenation of the IMSI and session ID.

Instead, we want to utilize our new StateIndexer interface and plugin element to generate this mapping asynchronously.

The StateIndexer type is intended to be run in a goroutine from the state service on new state updates (next diff).

## Changes to directoryd
To support routing all reads/writes of derived state through the directoryd service [0], we add getters/setters for the following derived-state mappings
- `{hwid -> hostname}`
- `{session ID -> IMSI}`

To logically collect location record manipulation, we move the getters for the following primary-state mappings just to the client API
- `{imsi -> session ID}`
- `{imsi -> hwid}`

## Changes to SyncRPC
This last `{imsi -> hwid}` mapping is used by SyncRPC, so we move the functionality back to directoryd.

## FeG use-case overview

{F230034335}

## Footnotes
[0] Rather than having to reconcile the indexer and servicer views of directoryd's storage

Reviewed By: xjtian

Differential Revision: D20293752

fbshipit-source-id: 7d62b5c0a6ae9d0e464639170141bdddc6f1a2fc