Enumerated type should not be defined as uint32.
Summary: QCI AVP is defined of type Enumerated in Diameter Gx interface as well as in policydb.proto file and not as unsigned integer. This change makes it consistent similar to the usage in "S6a_definitions.go". This will also rule out this type discrepancy as the potential cause of RAR decoding of non-zero QCI values as QCI_0.

Reviewed By: uri200

Differential Revision: D20454022

fbshipit-source-id: 1c6e893690478f0003fc9e47e6bb3aeec8aed6d4