create Network Hub tab with create service page (#428)
Summary:
Pull Request resolved: https://github.com/facebookincubator/symphony/pull/428

Pull Request resolved: https://github.com/facebookexternal/fbc/pull/2059

Create a new "Network Hub" tab in the Inventory Symphony project. Create a simple form for the "create service" page. This is a placeholder, and later the form will include more options and actually do stuff. This is fine to include in Symphony now because we need to explicitly enable the tab for a tenant to see it, so it won't interfere with current tenants.

Reviewed By: dlvhdr

Differential Revision: D20522835

fbshipit-source-id: 1fc25e3080307e991cc192d6c5d555c8c0eb182e