OVS: Add support for white spaces in command line argments.
Summary:
Currently OVS does not allow any space in arguments.
That forces us to add a hack to replace space with '_'. This patch make it cleaner to handle it.
User needs to provide double quotes to ovs-ofctl command parameters with white space. APN_NAME field does not store double quotes in the buffer. It is added in dump-flow output if APN_NAME has spaces in it.

Reviewed By: koolzz

Differential Revision: D20465438

fbshipit-source-id: 02f9b71f3225a7f4004bdb05f9c82575f69ff4c4