Add "Dynamic Strategy" as a passport strategy
Summary:
The OIDC strategy that comes with openid-client requires some attributes that
we only have access to at request time.  To work around that (so we dont have to
maintain our own strategy), I'm adding this DynamicStrategy which makes it
so we just create the substrategy at runtime.  It's kind of hackish, but it's
preferable to maintaining our own fork for multitenancy

Reviewed By: murtadha

Differential Revision: D20469571

fbshipit-source-id: 20e4895c389a98cd3c2be08de5f2a05b10260950