Removing check for maximum number of connected eNodeBs
Summary:
The MME sends an S1SetupFailure if the number of connected eNodeBs exceeds the
max_enbs (MAXENB in mme.conf.template). This is an inaccurate check as the
max_enbs value only controls the number of lists maintained in internal
hashtable for eNodeB descriptors. It does not limit the actual size of
the hashtable. This change removes the above logic.

Reviewed By: talkhasib

Differential Revision: D20630121

fbshipit-source-id: 3c6b0e17e0c68c1f1d510d295b39145d89ef9753