Add option to only run Gx tests, Gy tests, Authentication tests (#1391)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1391

Since we now have a lot of tests and it's not always useful to run all of the tests every time, I've added subcommands to the fabfile so we can run Auth/Gx/Gy tests separately.

This will be helpful when we start writing/testing Gy so we don't have to run the Gx tests every time.

When the param is not specified, it will just run all the tests.
CI will still continue to run all the tests as well.

Reviewed By: karthiksubraveti

Differential Revision: D20710633

fbshipit-source-id: 75ab79952f71ec919e0aaefcf5a1f31073b5a463