Fix gateways statuses
Show green circle for gateway when the enodeB is connected and MME is
connected.
Show yellow circle when RF transmit is disabled, if the enodeB is
disconnected or if the MME is disconnected - this indicates a fault with
the radio configuration which will not impact the site (since transmit
is disabled)
Show gray circle  when backhaul is down.
Show red circle in any other scenario when RF is enabled.