Session Manager's SessionStore supports deletion of sessions
Summary: This revision adds supports deletion of sessions to the `SessionStore` interface. Of note is the grouping of the deletion functionality into the `update_session` method. The reason for this is that each method of the interface is intended to appear atomic to the rest of the `sessiond` service, and as updates & deletions are done at the same time when credit usage is reported, then it makes sense to group this together.

Reviewed By: themarwhal

Differential Revision: D20501223

fbshipit-source-id: eda3338c0383fa12da9638deb22f47919da656a5