Session Manager's SessionStore supports deletion of sessions (#1407)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1407

Pull Request resolved: https://github.com/facebookexternal/fbc/pull/2144

This revision adds supports deletion of sessions to the `SessionStore` interface. Of note is the grouping of the deletion functionality into the `update_session` method. The reason for this is that each method of the interface is intended to appear atomic to the rest of the `sessiond` service, and as updates & deletions are done at the same time when credit usage is reported, then it makes sense to group this together.

Reviewed By: themarwhal

Differential Revision: D20501223

fbshipit-source-id: c59cc4dbf5094e4ef2e7746cf2cb0050b7fbea6b