Implement initial cwf operator to control cwf helm deployments
Summary:
This diff implements the initial cwf operator and it's controller. Currently
this controller is cwf helm deployments and health montoring
In a subsequent diff, the reconcilation logic for the HACluster controller will be extended to implement active/standby logic for CWAGs.

Most of the files provided below are the output of `operator-sdk generate ...`.

Reviewed By: xjtian

Differential Revision: D20682962

fbshipit-source-id: de182be219af6c3b6459a9c3bd5818a5d5d2406e