Add active/standby logic to CWF operator (#1432)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1432

This diff implements the main reconcilation logic needed for
the CWF Operator. This works by tracking the current active
gateway. Each reconcilation each we fetch health status.

If the active is unhealthy and the standby is healthy, we
trigger a failover. As a part of this, we track initialization
status, to ensure that the new active is in a proper state
(i.e sessions reloaded).

The actual implementation of obtaining health status
comes from a future diff that implements a local health
services. This diff will be updated with parsing that
status after that diff is finished.

Reviewed By: xjtian

Differential Revision: D20781307

fbshipit-source-id: 4cc756200e081927e7d8a75dc049b219935364a3