Adding S1ap integration tests to Magma docusaurus site
Summary:
This change adds a link to the "S1ap integration tests" page in the side bar of
Magma docusaurus website. It also adds a note in the "Quick start guide" page
for developers interested in LTE features. In most cases, developers focused on
EPC development/testing can skip the Orchestrator setup and proceed with a
VM-only dev setup and run S1ap tests.

Reviewed By: ardzoht

Differential Revision: D20822124

fbshipit-source-id: 2c5203f29dc4a763d738276dfa16ae6654297e69