add react-relay to hub ui (#1457) Updated with patch:
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1457

Add react-relay to the hub ui page and execute a simple "get version" query against the hub graph server.

For now the hub graph server is on localhost, eventually it will be deployed in the cluster and I'll change the code to reflect that.

Since we're using federation I've chosen to copy over the schema from introspecting the federated server. One option we discussed is concatenating the schema files but this ends up being pretty hard since they're not all in the same folder - there are different schema folder / files for every service that is being federated. Later we'll make a more robust solution for this.

Reviewed By: dlvhdr

Differential Revision: D20850294

fbshipit-source-id: 1bbc59bcfff78e68d8e920763163d14b2ea0f915
Signed-off-by: Jared Mullane <jmullane@fb.com>