Add secrets subchart to allow gwinfo to be mounted as secrets
Summary:
This diff adds a secrets subchart to the cwf helm chart. This allows
us to provide persistent gwinfo as a secret which will be mounted on pod recreation.
This ensure that when a pod is recreated, it'll continue to operate correctly
with the orchestrator.

Secrets can either be created independently or created as part of the helm
installation.

Reviewed By: xjtian

Differential Revision: D20906236

fbshipit-source-id: 7d97f1dca1699b6bc83f5914523e9e2ce732300b