Add secrets subchart to allow gwinfo to be mounted as secrets (#1460)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1460

This diff adds a secrets subchart to the cwf helm chart. This allows
us to provide persistent gwinfo as a secret which will be mounted on pod recreation.
This ensure that when a pod is recreated, it'll continue to operate correctly
with the orchestrator.

Secrets can either be created independently or created as part of the helm
installation.

Reviewed By: xjtian

Differential Revision: D20906236

fbshipit-source-id: fbd0b6fd9745ff3253621f796e426a8f23d2fb30