Fix LocalEnforcer's handle_cwf_roaming method by updating UpdateCriteria
Summary: Methods which update `SessionMap` should also update an UpdateCriteria so that writes back to `SessionStore` can be resolved correctly. In this revision, `LocalEnforcer`'s `handle_cwf_roaming` method is updated to an accept and modify a passed in UpdateCriteria so that session recycling will lead to a session being updated in the `SessionStore`.

Reviewed By: themarwhal

Differential Revision: D20800595

fbshipit-source-id: ba23b63b04ba397bc097f721a7bcfd415dbdf9bd