Bug fix for recycled sessions the sessiond. (#1428)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1428

Sessiond recycles sessions if a session exists with the right WLAN RAT type or exact configuration. When the session is in a transitional state to terminate, it creates a bug where session is deleted while MME thinks that the session still exists. This change does not recycle sessions in transitional state.

Reviewed By: themarwhal

Differential Revision: D20787659

fbshipit-source-id: e4c0a020ce9de4f94bf9f4c376b08be60e81a759