Add secrets subchart to FeG helmchart to allow persistent gwinfo
Summary:
This diff matches a previous diff enabling the same functionality
for the cwf helm chart. Here we are able to create a snowflake and gw_challenge.key
as a secret. This ensures that when a pod is recreated it will be able to continue
to function with the Orchestrator.

Reviewed By: xjtian

Differential Revision: D20966558

fbshipit-source-id: e03c22b1eed6643e81744b42752b94e060d8156f