Increase resource limits for orc8r grafana
Summary:
Grafana was crashlooping on staging. The logs provide 0 information so the only thing that makes sense is a lack of resources as per https://github.com/linkerd/linkerd2/issues/1747.

I'm setting these resources as the defaults on the chart values since we don't want users to run into this when they deploy.

Reviewed By: xjtian

Differential Revision: D21024104

fbshipit-source-id: cb11cee491483451c721653f5dbbc47a95ffd48a