Add dialect option for sequelize orm
Summary:
To more align with Magma technical requirements, we need to support mariadb rather than just mysql.

Sequelize added support for mariadb dialect in v5, and this diff just adds support to choose the dialect as an envioronment variable. Additionally, we now use a mariadb image in the development docker-compose file.

Differential Revision: D21077649

fbshipit-source-id: e41aa523973e3182e12993965d469758b33dbe74