Track policy rule lifetimes for subscribers
Summary:
Policy rule lifetimes need to be tracked by `session_manager` so that on restart of the service, rules can be correctly installed and uninstalled in the future if they were scheduled to do so.

This revision does not include the rescheduling of rule activation/deactivation upon service restart. That will come in a separate revision

## Changes
- Add a `RuleLifetime` struct to more easily track policy rule lifetimes
- `SessionState` tracks policy rule lifetimes, both for rules already installed, and those scheduled to be installed
- `SessionState` methods modified so that any rule activations and tracking require also specifying the lifetime of the policy rule
- Refactored `LocalEnforcer` to remove `process_create_session_response`, decomposing its functionality into a new method, `filter_rule_installs` and an existing method, `process_rule_installs`

Differential Revision: D21055817

fbshipit-source-id: c53335aeb4102b805f486e71dfc94675d8a076d0