Include "Template" dashboard for nms grafana
Summary: The `networkID` and `gatewayID` variables are fairly difficult to set up correctly, and a user would have a hard time figuring out how to do this if they wanted that feature in a custom dashboard. By providing a template which has a single graph (for variable usage demo), users can simply copy this template then build their dashboard off of this base.

Reviewed By: rckclmbr

Differential Revision: D21145359

fbshipit-source-id: 32befb30c0b1a6d012ebe3f701f676c78f50960d