Add stop functionality to GRE probe for CWAG health service (#1557)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1557

When a CWAG is demoted, we block ICMP on the eth1
interface to ensure that traffic will properly be re-routed.
Due to this, the ICMP GRE probe will always fail when a gateway
is in standby. This causes the gateway to appear unhealthy.
To avoid this, we stop the probe on gateway demotion and start
the probe on gateway promotion.

Reviewed By: xjtian

Differential Revision: D21219399

fbshipit-source-id: 60e2a4a591710f99f5a72167c3d4d5008cc89832