Add load criteria to Search in support of ListKeysByNetwork.
Summary:
## Rationale

State indexers' reindex method needs to pass all known state, independent of network, through the indexer to be reindexed. We currently have no way to get a list of all state/state IDs in a cross-network context. `ListKeysByNetwork` us the hook to get cross-network state IDs, which can then be converted to state on a per-network basis via the normal methods.

## Implementation

The easiest way to get all keys is to use the existing `Search` method, so we
- update `Search` to (a) allow nil network filters and (b) function correctly when all filters are nil (return all blobs)
- update `Search` to support a load criteria -- in this case, whether to load blob values
- add a default method `ListKeysByNetwork` as a wrapper around `Search` with nil filters and a "no values" load criteria

Reviewed By: xjtian

Differential Revision: D21027151

fbshipit-source-id: 8e66ce61bfcfebe9475bfbfd6fa897a583f3789f