Multiple diameter (integ_test) - Test to check UE control data flows through the right PCRF and OCS (#1553)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1553

This is a basic tests with the new set up of  2 PCRF and 2 OCS.
The UE gets configured only in one of those PCRF and OCS instances based on its IMSI.
The test consists on single UE sending 500k, reporting to PCRF and OCS and detaching.
The objective is to make sure the data flows through the right instances.

Some generic issues that may need further investigation:
- For some reason OCS and PCRF seems to get the subscribes accounts deleted even before we send the detach. This prevents the credit to be reported properly
- UESim uses only one MAC address, so it is not possible to send traffic from more than one UE. So any tests is currently limited to one single UE in data plane

Reviewed By: themarwhal

Differential Revision: D21147346

fbshipit-source-id: e04dd96bd59e4f9b290dc35a87ac478f7c64b371