Re-add sessions to AAA server on stateless sessiond restart (#1586)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1586

This diff adds logic to re-create AAA-server sessions on sessiond restart.
This is only used for CWF sessions when sessiond is running with stateless configured.
After a failover, sessiond is restarted. This restart will trigger a re-creation of flows
in pipelined and now a re-creation of AAA server sessions.

Reviewed By: themarwhal

Differential Revision: D21297546

fbshipit-source-id: ec382f372b13cbc10896f2e802ee78ab6edd4d7b