Fix redis disconnect bug for RedisStoreClient in sessiond
Summary:
This diff attempts to reconnect to Redis before reading or writing
if the Redis client is unconnected. This should fix an issue where the redis
client gets disconnected but redis is still available. This was happening often,
causing sessiond do unnecessarily crash loop.

Sessiond will still crash if Redis is unavailable.

Reviewed By: andreilee

Differential Revision: D21317580

fbshipit-source-id: 3a457a8ab5446d0c717e014b922aa2f946409e6b