Only run setup_aaa_sessions when sessiond is configured for stateless CWF
Summary:
This diff fixes a bug causing sessiond to crash when not
configured for carrier wifi support. This is due to the AAA client
thread not running. This is fixed by updating the logic to only run
the setup when stateless and configured for CWF.

It also ensures that this request is only sent with non-empty sessions.

Reviewed By: ardzoht

Differential Revision: D21391738

fbshipit-source-id: 15d1dc288771a57d6059b0f75fca221699dfe52f