Dedicated bearer context removal and GTP tunnel deletion when default bearer is deactivated
Summary:
When network initiated bearer deactivation is executed for the default bearer, the context for the dedicated bearers in the same session must be also deleted and GTP tunnel rules must be removed. This was not done properly and the behavior is corrected with this diff.

Diff fixes memory leak as well as duplicate tunnel rule deletion for default bearer (it is deleted once in the modified function and second in function call ```sgw_handle_sgi_endpoint_deleted(&sgi_delete_end_point_request, imsi64);```).

Reviewed By: ssanadhya

Differential Revision: D21421449

fbshipit-source-id: 1de3713b6fb3232326e1a22d075c6cbb6fe82f91