Schedule termination for sessions that are awaiting termination on restart (#1634)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1634

## Background on the SESSION_TERMINATION_SCHEDULED state
This state is only used for CWF currently. A session goes into this state when a session is initialized without a valid wallet. This usually means that the PCRF will not send down PCC rules to browse internet, and there is only the Zero-Rated rule installed. (ZR rule only allows access to a data purchasing site). We do not immediately reject the session creation in this scenario, since we want to allow the user to purchase data. Instead, we will schedule a termination in a configured amount of time. (The default is 30 seconds).

## On SessionD Restart
When SessionD restarts and we re-load all sessions, we want to make sure that the user gets kicked off in some time. We can keep this logic dumb and always just use the configured kick off time. (So if a user gets scheduled to terminate at time T+30 and SessionD restarts at time T+15, the user will end up getting kicked off at time T+45. We can change this if we need to, for now this should be good enough.

## Some Bug Fix
I realized when writing the unit test for this feature that I introduced a bug a few diffs ago. In `terminate_service` we should ignore the termination only if the session is already terminating, and not "not active". This is because "not active" also includes SESSION_TERMINATION_SCHEDULED.

## Summary of Changes
- Add 'schedule_termination' to consolidate the scheduling logic
- Add 'is_terminating' and 'get_state' to SessionState
- In `sync_sessions_on_restart` schedule terminations for sessions with SESSION_TERMINATION_SCHEDULED
- Add unit test

Reviewed By: andreilee

Differential Revision: D21450420

fbshipit-source-id: 7dc04d22149ac395b3d077552f330e07537f9949