Add a null case test for sync_sessions_on_restart
Summary:
`sync_sessions_on_restart` can add events on the event loop but was called before the event loop was attached to local enforcer. This can cause a sessiond crash when things are added to the event loop before initialization.
changing that here so that it's called after the attach.

Differential Revision: D21482296

fbshipit-source-id: 1d755b278cef7c9c997b7929587c3c532a69312b