Fix sessiondcrash on sync_sessions_on_restart (#1643)
Summary:
Pull Request resolved: https://github.com/facebookincubator/magma/pull/1643

`sync_sessions_on_restart` can add events on the event loop but was called before the event loop was attached to local enforcer. This can cause a sessiond crash when things are added to the event loop before initialization.
changing that here so that it's called after the attach.

Reviewed By: ardzoht, koolzz

Differential Revision: D21482296

fbshipit-source-id: d02533425cac14d0e5a21a07b0a1b9a547178f47